refactor: (ifo): Change get lp modal name and make it generic
