+++ src/config/localization/translations.json
@@ -1757,5 +1757,7 @@
   "Failed to load price chart for this pair": "Failed to load price chart for this pair",
   "You can swap WBNB for BNB (and vice versa) with no trading fees.": "You can swap WBNB for BNB (and vice versa) with no trading fees.",
   "Exchange rate is always 1 to 1.": "Exchange rate is always 1 to 1.",
-  "You can contribute to one or both, it doesn’t matter: only your overall contribution is counted for the achievement.": "You can contribute to one or both, it doesn’t matter: only your overall contribution is counted for the achievement."
+  "You can contribute to one or both, it doesn’t matter: only your overall contribution is counted for the achievement.": "You can contribute to one or both, it doesn’t matter: only your overall contribution is counted for the achievement.",
+  "You’ll need %symbol% tokens to participate in the IFO!": "You’ll need %symbol% tokens to participate in the IFO!",
+  "Get %symbol%, or make sure your tokens aren’t staked somewhere else.": "Get %symbol%, or make sure your tokens aren’t staked somewhere else."
 }
