--- src/views/Ifos/components/IfoFoldableCard/IfoPoolCard/GetLpModal.tsx
@@ -1,39 +1,0 @@
-import React from 'react'
-import { Modal, ModalBody, Text, Image, Button, Link, OpenNewIcon } from '@pancakeswap/uikit'
-import { Token } from '@pancakeswap/sdk'
-import { useTranslation } from 'contexts/Localization'
-
-interface Props {
-  currency: Token
-  onDismiss?: () => void
-}
-
-const GetLpModal: React.FC<Partial<Props>> = ({ currency, onDismiss }) => {
-  const { t } = useTranslation()
-  return (
-    <Modal title={t('CAKE required')} onDismiss={onDismiss}>
-      <ModalBody maxWidth="288px">
-        <Image
-          src={`/images/${currency.symbol.split(' ')[0].toLocaleLowerCase()}.svg`}
-          width={72}
-          height={72}
-          margin="auto"
-          mb="24px"
-        />
-        <Text mb="16px">{t('You’ll need CAKE tokens to participate in the IFO!')}</Text>
-        <Text mb="24px">{t('Get CAKE, or make sure your tokens aren’t staked somewhere else.')}</Text>
-        <Button
-          as={Link}
-          external
-          href="/swap?outputCurrency=0x0e09fabb73bd3ade0a17ecc321fd13a19e81ce82"
-          endIcon={<OpenNewIcon color="white" />}
-          minWidth="100%" // Bypass the width="fit-content" on Links
-        >
-          {t('Get %symbol%', { symbol: 'CAKE' })}
-        </Button>
-      </ModalBody>
-    </Modal>
-  )
-}
-
-export default GetLpModal
